refactor(aluminium): clean up patches
