+++ Ignitron.Aluminium/Events/ClientLoopEvents.Patches.cs
